Incorporando Editar VTA

+++ Dominio/Especialista.cs
@@ -15,8 +15,8 @@ namespace Dominio
         public string Telefono { get; set; }
         public DateTime Cumple { get; set; }
 
-        public int Sueldo { get; set; }
-        public int Porcentaje { get; set; }
+        public Decimal Sueldo { get; set; }
+        public double Porcentaje { get; set; }
 
 
         public override string ToString()
